tests: remove -c from pascal tests
tanco's -c flag doesn't seem to work right on linux.
but on windows, you need it when the interpreter to run
(lilt, node, etc) isn't in the current directory. I will
have to fix this for the other languages, but with the
pascal version, the executable is always in pas/ so you
don't need it.